+++ src/hooks/players/players.hook.ts
@@ -6,10 +6,11 @@ import { useQuery } from "@tanstack/react-query";
 import { useEffect } from "react";
 
 export function useFetchPlayers() {
+  const access_token = useSolidStore(selectAccessToken);
   const selectedTeamId = useSolidStore(selectSelectedTeamId);
   const setPlayers = useSolidStore((state) => state.setPlayers);
   const { data, isLoading, error } = useQuery(
-    playerWithStatsQueryOptions(selectedTeamId)
+    playerWithStatsQueryOptions(selectedTeamId, access_token)
   );
 
   useEffect(() => {
